# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Blog API",
    "version": "1.0.0",
    "description": "REST API for the Blog backend (Express 5, MongoDB/Mongoose, JWT). All JSON routes are under `/api/v1`.",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login/register response."
      },
      "cookieRefreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HTTP-only JWT refresh cookie set on login/register."
      }
    },
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ValidationError"
          },
          "errors": {
            "type": "object",
            "additionalProperties": true,
            "description": "express-validator mapped errors"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/comment/blog/{blogId}": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Add comment on blog",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blogId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created; blog `commentsCount` incremented",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comment": {
                      "type": "object"
                    },
                    "commentsCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Blog not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "List comments for a blog",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "blogId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of comments (key `comment` in JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comment": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Blog not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/comment/blog/{commentId}": {
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comment Mongo ObjectId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted (no body)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not comment owner and not admin"
          },
          "404": {
            "description": "Comment or blog not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}